refactor: 모듈명 변경
